Move the grading library into the eval layer
Graders, Scorecard, the judge runner, and the grading-module loader are
general-purpose scoring code that happened to live under optimize/. Moving them
to lib/eval/grading/ lets eval reach them without importing upward into
optimize, which is the dependency direction we want.

Pure relocation: no signature or behavior changes, only import paths. The 414
tests in lib/eval, lib/optimize, and lib/cli/eval pass with no test-body edits,
which is the proof.
