+++ packages/agency-lang/lib/optimize/baseOptimizer.workdir.test.ts
@@ -4,9 +4,9 @@ import * as path from "path";
 
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { BaseGrader } from "./grading/baseGrader.js";
-import type { Grade, GraderInput, GraderOptions, Input } from "./grading/types.js";
-import type { Scorecard } from "./grading/scorecard.js";
+import { BaseGrader } from "@/eval/grading/baseGrader.js";
+import type { Grade, GraderInput, GraderOptions, Input } from "@/eval/grading/types.js";
+import type { Scorecard } from "@/eval/grading/scorecard.js";
 import { BaseOptimizer } from "./baseOptimizer.js";
 import type { OptimizeTargetSet } from "./targets.js";
 import type { OptimizeResult } from "./types.js";
