+++ packages/agency-lang/lib/optimize/optimizers/gepa.test.ts
@@ -5,8 +5,8 @@ import * as path from "path";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { Gepa, type GepaConfig, type GepaDeps } from "./gepa.js";
-import { BaseGrader } from "../grading/baseGrader.js";
-import type { Grade, GraderInput } from "../grading/types.js";
+import { BaseGrader } from "@/eval/grading/baseGrader.js";
+import type { Grade, GraderInput } from "@/eval/grading/types.js";
 import type { RunInput } from "../baseOptimizer.js";
 import type { OptimizeMutationPreview } from "../sourceMutator.js";
 import type { OptimizeTargetSet } from "../targets.js";
